Windows: Fixed warning in `windows/fs.hpp`.
Removes an implicit cast from `ULONGLONG` to `double`.

Review: https://reviews.apache.org/r/52207/

+++ 3rdparty/stout/include/stout/windows/fs.hpp
@@ -69,7 +69,7 @@ inline Try<double> usage(const std::string& path = "/")
         "Error invoking 'GetDiskFreeSpaceEx' on '" + path + "'");
   }
 
-  double used = total_bytes.QuadPart - free_bytes.QuadPart;
+  double used = static_cast<double>(total_bytes.QuadPart - free_bytes.QuadPart);
   return used / total_bytes.QuadPart;
 }
